requirements.txt
git+https://github.com/agronholm/anyio
tgcrypto
aiohttp
pymongo[srv]==4.3.3
marshmallow==3.19.0
motor==3.1.1
umongo==3.1.0
requests
bs4
git+https://github.com/Joelkb/cinemagoer 
datetime
psutil
python-dotenv
pyrofork==2.3.43
